Add files via upload
Please note that the build folder of this android studio project has been deleted in order to reduce the size. Hence you cannot import the project as such to work in android studio, However you could copy the 'src' folder and merge with your existing 'src' [by creating a new project in android studio] and rebuild the app to see it work.